+++ src/secure_string_cipher/__init__.py
@@ -26,7 +26,7 @@ from .timing_safe import (
 )
 from .utils import ProgressBar, colorize, handle_timeout, secure_overwrite
 
-__version__ = "1.0.22"
+__version__ = "1.0.23"
 __author__ = "TheRedTower"
 __email__ = "security@avondenecloud.uk"